requirements.txt
requests==2.25.1
netaddr==0.8.0
cryptography==3.4.7
ldap3==2.9
Flask==1.1.2
validators==0.18.2
ntplib==0.3.4
dnspython==2.1.0
waitress==2.0.0
